move header and auth
