+++ basic/main.wasp
@@ -14,7 +14,7 @@ app __waspAppName__ {
           name: "Basic App",
           email: "hello@example.com"
         },
-        userSignupFields: import { userSignupFields } from "@src/auth/email",
+        userSignupFields: import { userSignupFields } from "@src/auth/email/userSignupFields",
         emailVerification: {
           clientRoute: EmailVerificationRoute,
         },
@@ -42,27 +42,27 @@ page TaskDashboardPage {
 
 route LoginRoute { path: "/login", to: LoginPage }
 page LoginPage {
-  component: import { LoginPage } from "@src/auth/LoginPage.tsx"
+  component: import { LoginPage } from "@src/auth/email/LoginPage.tsx"
 }
 
 route SignupRoute { path: "/signup", to: SignupPage }
 page SignupPage {
-  component: import { SignupPage } from "@src/auth/SignupPage.tsx"
+  component: import { SignupPage } from "@src/auth/email/SignupPage.tsx"
 }
 
 route RequestPasswordResetRoute { path: "/request-password-reset", to: RequestPasswordResetPage }
 page RequestPasswordResetPage {
-  component: import { RequestPasswordResetPage } from "@src/auth/RequestPasswordResetPage.tsx",
+  component: import { RequestPasswordResetPage } from "@src/auth/email/RequestPasswordResetPage.tsx",
 }
 
 route PasswordResetRoute { path: "/password-reset", to: PasswordResetPage }
 page PasswordResetPage {
-  component: import { PasswordResetPage } from "@src/auth/PasswordResetPage.tsx",
+  component: import { PasswordResetPage } from "@src/auth/email/PasswordResetPage.tsx",
 }
 
 route EmailVerificationRoute { path: "/email-verification", to: EmailVerificationPage }
 page EmailVerificationPage {
-  component: import { EmailVerificationPage } from "@src/auth/EmailVerificationPage.tsx",
+  component: import { EmailVerificationPage } from "@src/auth/email/EmailVerificationPage.tsx",
 }
 
 query getTasks {
