+++ basic/src/auth/email/EmailVerificationPage.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import { VerifyEmailForm } from "wasp/client/auth";
-import { AuthLayout } from "./components/AuthLayout";
+import { AuthLayout } from "../AuthLayout";
 
 export function EmailVerificationPage() {
   return (
