+++ basic/src/auth/email/RequestPasswordResetPage.tsx
@@ -1,5 +1,5 @@
 import { ForgotPasswordForm } from "wasp/client/auth";
-import { AuthLayout } from "./components/AuthLayout";
+import { AuthLayout } from "../AuthLayout";
 
 export function RequestPasswordResetPage() {
   return (
