+++ basic/src/auth/email/SignupPage.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import { SignupForm } from "wasp/client/auth";
-import { AuthLayout } from "./components/AuthLayout";
+import { AuthLayout } from "../AuthLayout";
 
 export function SignupPage() {
   return (
